@@ -28,19 +28,21 @@
         }
     ],
     "require": {
-        "php": "^7.2",
-        "guzzlehttp/guzzle": "~6.3.0 || ~6.4.0",
-        "illuminate/cache": "^5.8.2|^6.0.0",
-        "illuminate/container": "^5.8.2|^6.0.0",
-        "illuminate/config": "^5.8.2|^6.0.0",
-        "illuminate/filesystem": "^5.8.2|^6.0.0",
-        "illuminate/redis": "^5.8.2|^6.0.0",
-        "illuminate/database": "^5.8.2|^6.0.0",
+        "php": "^8.0",
+        "guzzlehttp/guzzle": "^7.3",
+        "illuminate/cache": "^8",
+        "illuminate/container": "^8",
+        "illuminate/config": "^8",
+        "illuminate/filesystem": "^8",
+        "illuminate/redis": "^8",
+        "illuminate/database": "^8",
         "predis/predis": "^1.1",
-        "symfony/http-kernel": "~3|~4"
+        "symfony/http-kernel": "^5"
     },
     "require-dev": {
-        "phpunit/phpunit": "^8.0.4"
+        "phpunit/phpunit": "^9",
+        "mockery/mockery": "^1",
+        "slope-it/clock-mock": "^0.2.0"
     },
     "autoload": {
         "psr-4": {
