@@ -2,36 +2,26 @@
 
 namespace hamburgscleanest\GuzzleAdvancedThrottle\Cache\Adapters;
 
-use DateInterval;
-use DateTime;
 use GuzzleHttp\Psr7\Response;
 use hamburgscleanest\GuzzleAdvancedThrottle\Cache\CachedResponse;
 use hamburgscleanest\GuzzleAdvancedThrottle\RequestInfo;
+use hamburgscleanest\GuzzleAdvancedThrottle\SystemClock;
+use hamburgscleanest\GuzzleAdvancedThrottle\TimeKeeper;
 use Psr\Http\Message\ResponseInterface;
 use Illuminate\Config\Repository;
 
-/**
- * Class ArrayAdapter
- * @package hamburgscleanest\GuzzleAdvancedThrottle\Cache\Adapters
- */
 class ArrayAdapter extends BaseAdapter
 {
-
     /** @var string */
     private const RESPONSE_KEY = 'response';
     /** @var string */
     private const EXPIRATION_KEY = 'expires_at';
-    /** @var array */
-    private $_storage = [];
 
-    /**
-     * StorageInterface constructor.
-     * @param Repository|null $config
-     */
+    private array $_storage = [];
+
     public function __construct(?Repository $config = null)
     {
-        if ($config === null)
-        {
+        if ($config === null) {
             return;
         }
 
@@ -39,77 +29,49 @@ class ArrayAdapter extends BaseAdapter
         $this->_allowEmptyValues = $config->get('cache.allow_empty', $this->_allowEmptyValues);
     }
 
-    /**
-     * @param string $host
-     * @param string $key
-     * @param int $requestCount
-     * @param \DateTime $expiresAt
-     * @param int $remainingSeconds
-     */
-    public function save(string $host, string $key, int $requestCount, DateTime $expiresAt, int $remainingSeconds) : void
+    public function save(string $host, string $key, int $requestCount, TimeKeeper $timeKeeper): void
     {
-        $this->_storage[$host][$key] = RequestInfo::create($requestCount, $expiresAt->getTimestamp(), $remainingSeconds);
+        $expiration = $timeKeeper->getExpiration();
+        if ($expiration === null) {
+            unset($this->_storage[$host][$key]);
+            return;
+        }
+
+        $this->_storage[$host][$key] = RequestInfo::create(
+            $requestCount,
+            $expiration->getTimestamp(),
+            $timeKeeper->getRemainingSeconds()
+        );
     }
 
-    /**
-     * @param string $host
-     * @param string $key
-     * @return RequestInfo|null
-     */
-    public function get(string $host, string $key) : ?RequestInfo
+    public function get(string $host, string $key): ?RequestInfo
     {
         return $this->_storage[$host][$key] ?? null;
     }
 
-    /**
-     * @param ResponseInterface $response
-     * @param string $host
-     * @param string $path
-     * @param string $key
-     * @param int $expiresAt
-     * @throws \Exception
-     */
-    protected function _saveResponse(ResponseInterface $response, string $host, string $path, string $key) : void
+    protected function _saveResponse(ResponseInterface $response, string $host, string $path, string $key): void
     {
         $this->_storage[self::STORAGE_KEY][$host][$path][$key] = [
             self::RESPONSE_KEY   => new CachedResponse($response),
-            self::EXPIRATION_KEY => (new DateTime())->add(new DateInterval('PT' . $this->_ttl . 'M'))->getTimestamp()
+            self::EXPIRATION_KEY => SystemClock::create()->advanceMinutes($this->_ttl)->now()->getTimestamp()
         ];
     }
 
-    /**
-     * @param string $host
-     * @param string $path
-     * @param string $key
-     * @return null|Response
-     */
-    protected function _getResponse(string $host, string $path, string $key) : ?Response
+    protected function _getResponse(string $host, string $path, string $key): ?Response
     {
         $response = $this->_storage[self::STORAGE_KEY][$host][$path][$key] ?? null;
 
-        if ($response !== null)
-        {
-            if ($response[self::EXPIRATION_KEY] > \time())
-            {
+        if ($response !== null) {
+            if ($response[self::EXPIRATION_KEY] > \time()) {
                 /** @var CachedResponse|null $cachedResponse */
                 $cachedResponse = $response[self::RESPONSE_KEY];
 
                 return $cachedResponse ? $cachedResponse->getResponse() : null;
             }
 
-            $this->_invalidate($host, $path, $key);
+            unset($this->_storage[self::STORAGE_KEY][$host][$path][$key]);
         }
 
         return null;
-    }
-
-    /**
-     * @param string $host
-     * @param string $path
-     * @param string $key
-     */
-    private function _invalidate(string $host, string $path, string $key) : void
-    {
-        unset($this->_storage[self::STORAGE_KEY][$host][$path][$key]);
     }
 }
