@@ -7,21 +7,9 @@ use GuzzleHttp\Promise\PromiseInterface;
 use Psr\Http\Message\RequestInterface;
 use Symfony\Component\HttpKernel\Exception\TooManyRequestsHttpException;
 
-/**
- * Class Cache
- * @package hamburgscleanest\GuzzleAdvancedThrottle\Cache\Strategies
- */
 class Cache extends Cacheable
 {
-
-    /**
-     * @param RequestInterface $request
-     * @param callable $handler
-     * @return PromiseInterface
-     * @throws \InvalidArgumentException
-     * @throws \Symfony\Component\HttpKernel\Exception\TooManyRequestsHttpException
-     */
-    public function request(RequestInterface $request, callable $handler) : PromiseInterface
+    public function request(RequestInterface $request, callable $handler): PromiseInterface
     {
         try
         {
@@ -30,8 +18,7 @@ class Cache extends Cacheable
         catch (TooManyRequestsHttpException $tooManyRequestsHttpException)
         {
             $response = $this->_getResponse($request);
-            if ($response !== null)
-            {
+            if ($response !== null) {
                 return new FulfilledPromise($response);
             }
 
