@@ -6,39 +6,22 @@ use hamburgscleanest\GuzzleAdvancedThrottle\Cache\Interfaces\StorageInterface;
 use Psr\Http\Message\RequestInterface;
 use SplObjectStorage;
 
-/**
- * Class RequestLimitGroup
- * @package hamburgscleanest\GuzzleAdvancedThrottle
- */
 class RequestLimitGroup
 {
+    private SplObjectStorage $_requestLimiters;
+    private int $_retryAfter = 0;
 
-    /** @var SplObjectStorage */
-    private $_requestLimiters;
-
-    /** @var int */
-    private $_retryAfter = 0;
-
-    /**
-     * RequestLimitGroup constructor.
-     */
     public function __construct()
     {
         $this->_requestLimiters = new SplObjectStorage();
     }
 
-    /**
-     * @return RequestLimitGroup
-     */
-    public static function create() : self
+    public static function create(): self
     {
         return new static();
     }
 
-    /**
-     * @return int
-     */
-    public function getRetryAfter() : int
+    public function getRetryAfter(): int
     {
         return $this->_retryAfter;
     }
@@ -46,24 +29,18 @@ class RequestLimitGroup
     /**
      * We have to cycle through all the limiters (no early return).
      * The timers of each limiter have to be updated despite of another limiter already preventing the request.
-     *
-     * @param RequestInterface $request
-     * @param array $options
-     * @return bool
-     * @throws \Exception
      */
-    public function canRequest(RequestInterface $request, array $options = []) : bool
+    public function canRequest(RequestInterface $request, array $options = []): bool
     {
         $groupCanRequest = true;
         $this->_requestLimiters->rewind();
-        while ($this->_requestLimiters->valid())
-        {
+        while ($this->_requestLimiters->valid()) {
             /** @var RequestLimiter $requestLimiter */
             $requestLimiter = $this->_requestLimiters->current();
 
             $canRequest = $requestLimiter->canRequest($request, $options);
-            if ($groupCanRequest && !$canRequest)
-            {
+
+            if ($groupCanRequest && !$canRequest) {
                 $groupCanRequest = false;
                 $this->_retryAfter = $requestLimiter->getRemainingSeconds();
             }
@@ -74,46 +51,28 @@ class RequestLimitGroup
         return $groupCanRequest;
     }
 
-    /**
-     * @param RequestLimiter $requestLimiter
-     * @return RequestLimitGroup
-     */
-    public function addRequestLimiter(RequestLimiter $requestLimiter) : self
+    public function addRequestLimiter(RequestLimiter $requestLimiter): self
     {
         $this->_requestLimiters->attach($requestLimiter);
 
         return $this;
     }
 
-    /**
-     * @param RequestLimiter $requestLimiter
-     * @return RequestLimitGroup
-     */
-    public function removeRequestLimiter(RequestLimiter $requestLimiter) : self
+    public function removeRequestLimiter(RequestLimiter $requestLimiter): self
     {
         $this->_requestLimiters->detach($requestLimiter);
 
         return $this;
     }
 
-    /**
-     * @return int
-     */
-    public function getRequestLimiterCount() : int
+    public function getRequestLimiterCount(): int
     {
         return $this->_requestLimiters->count();
     }
 
-    /**
-     * @param string $host
-     * @param array $rules
-     * @param StorageInterface $storage
-     * @throws \Exception
-     */
-    public function addRules(string $host, array $rules, StorageInterface $storage) : void
+    public function addRules(string $host, array $rules, StorageInterface $storage): void
     {
-        foreach ($rules as $rule)
-        {
+        foreach ($rules as $rule) {
             $this->addRequestLimiter(
                 RequestLimiter::createFromRule(
                     $host,
