@@ -1,70 +1,34 @@
 <?php
-/**
- * Created by PhpStorm.
- * User: jporter
- * Date: 5/29/18
- * Time: 7:53 PM
- */
 
 namespace hamburgscleanest\GuzzleAdvancedThrottle\Tests;
 
-
-use DateTime;
 use hamburgscleanest\GuzzleAdvancedThrottle\Cache\Interfaces\StorageInterface;
 use hamburgscleanest\GuzzleAdvancedThrottle\RequestInfo;
+use hamburgscleanest\GuzzleAdvancedThrottle\TimeKeeper;
 use Illuminate\Config\Repository;
 use Psr\Http\Message\RequestInterface;
 use Psr\Http\Message\ResponseInterface;
 
-/**
- * Class DummyStorageAdapter
- * @package hamburgscleanest\GuzzleAdvancedThrottle\Tests
- */
 class DummyStorageAdapter implements StorageInterface
 {
-
-    /**
-     * StorageInterface constructor.
-     * @param Repository|null $config
-     */
     public function __construct(?Repository $config = null)
     {
     }
 
-    /**
-     * @param string $host
-     * @param string $key
-     * @param int $requestCount
-     * @param DateTime $expiresAt
-     * @param int $remainingSeconds
-     */
-    public function save(string $host, string $key, int $requestCount, DateTime $expiresAt, int $remainingSeconds) : void
+    public function save(string $host, string $key, int $requestCount, TimeKeeper $timeKeeper): void
     {
     }
 
-    /**
-     * @param string $host
-     * @param string $key
-     * @return RequestInfo|null
-     */
-    public function get(string $host, string $key) : ?RequestInfo
+    public function get(string $host, string $key): ?RequestInfo
     {
         return null;
     }
 
-    /**
-     * @param RequestInterface $request
-     * @param ResponseInterface $response
-     */
-    public function saveResponse(RequestInterface $request, ResponseInterface $response) : void
+    public function saveResponse(RequestInterface $request, ResponseInterface $response): void
     {
     }
 
-    /**
-     * @param RequestInterface $request
-     * @return ResponseInterface|null
-     */
-    public function getResponse(RequestInterface $request) : ?ResponseInterface
+    public function getResponse(RequestInterface $request): ?ResponseInterface
     {
         return null;
     }
