@@ -2,62 +2,51 @@
 
 namespace hamburgscleanest\GuzzleAdvancedThrottle\Tests;
 
-use DateTime;
 use GuzzleHttp\Psr7\Request;
 use GuzzleHttp\Psr7\Response;
 use hamburgscleanest\GuzzleAdvancedThrottle\Cache\Adapters\LaravelAdapter;
 use hamburgscleanest\GuzzleAdvancedThrottle\Exceptions\LaravelCacheConfigNotSetException;
+use hamburgscleanest\GuzzleAdvancedThrottle\TimeKeeper;
 use Illuminate\Config\Repository;
 use PHPUnit\Framework\TestCase;
 
-/**
- * Class LaravelAdapterTest
- * @package hamburgscleanest\GuzzleAdvancedThrottle\Tests
- */
 class LaravelAdapterTest extends TestCase
 {
-
     /** @test */
-    public function throws_an_exception_when_config_is_not_set() : void
+    public function throws_an_exception_when_config_is_not_set(): void
     {
         $this->expectException(LaravelCacheConfigNotSetException::class);
 
         new LaravelAdapter();
     }
 
-    /** @test
-     * @throws \Exception
-     */
-    public function stores_and_retrieves_data() : void
+    /** @test */
+    public function stores_and_retrieves_data(): void
     {
         $host = 'test';
         $key = 'my_key';
         $requestCount = 12;
-        $expiresAt = new DateTime();
-        $remainingSeconds = 120;
+
+        $timeKeeper = new TimeKeeper(120);
+        $timeKeeper->start();
 
         $laravelAdapter = new LaravelAdapter($this->_getConfig());
-        $laravelAdapter->save($host, $key, $requestCount, $expiresAt, $remainingSeconds);
+        $laravelAdapter->save($host, $key, $requestCount, $timeKeeper);
 
         $requestInfo = $laravelAdapter->get($host, $key);
         static::assertNotNull($requestInfo);
-        static::assertEquals($requestInfo->remainingSeconds, $remainingSeconds);
+        static::assertEquals($requestInfo->remainingSeconds, $timeKeeper->getRemainingSeconds());
         static::assertEquals($requestInfo->requestCount, $requestCount);
-        static::assertEquals($requestInfo->expiresAt->getTimestamp(), $expiresAt->getTimestamp());
+        static::assertEquals($requestInfo->expiresAt->format('Y-m-d H:i:s'), $timeKeeper->getExpiration()->format('Y-m-d H:i:s'));
     }
 
-    /**
-     * @return Repository
-     */
-    private function _getConfig() : Repository
+    private function _getConfig(): Repository
     {
         return new Repository(require 'config/app.php');
     }
 
-    /** @test
-     * @throws \Exception
-     */
-    public function stored_value_gets_invalidated_when_expired() : void
+    /** @test */
+    public function stored_value_gets_invalidated_when_expired(): void
     {
         $request = new Request('GET', 'www.test.com');
         $response = new Response(200, [], null);
@@ -72,11 +61,8 @@ class LaravelAdapterTest extends TestCase
         static::assertNull($storedResponse);
     }
 
-    /**
-     * @test
-     * @throws \Exception
-     */
-    public function does_not_store_empty_values() : void
+    /** @test */
+    public function does_not_store_empty_values(): void
     {
         $request = new Request('GET', 'www.test.com');
         $nullResponse = new Response(200, [], null);
@@ -96,10 +82,8 @@ class LaravelAdapterTest extends TestCase
         static::assertNull($laravelAdapter->getResponse($request));
     }
 
-    /**
-     * @test
-     */
-    public function stores_empty_values_when_allowed() : void
+    /** @test */
+    public function stores_empty_values_when_allowed(): void
     {
         $request = new Request('GET', 'www.test.com');
         $nullResponse = new Response(200, [], null);
@@ -117,6 +101,5 @@ class LaravelAdapterTest extends TestCase
         $laravelAdapter->saveResponse($request, $emptyResponse);
 
         static::assertEmpty((string) $laravelAdapter->getResponse($request)->getBody());
-
     }
 }
