@@ -13,27 +13,22 @@ use Illuminate\Config\Repository;
 use PHPUnit\Framework\TestCase;
 use Symfony\Component\HttpKernel\Exception\TooManyRequestsHttpException;
 
-/**
- * Class MemcachedDriverTest
- * @package hamburgscleanest\GuzzleAdvancedThrottle\Tests
- */
 class MemcachedDriverTest extends TestCase
 {
 
-    /** @test
-     * @throws \Exception
-     */
-    public function throws_servers_not_set_exception() : void
+    /** @test */
+    public function throws_servers_not_set_exception(): void
     {
         $this->expectException(MemcachedServersNotSetException::class);
 
-        new RequestLimitRuleset([
-            'www.test.de' => [
-                [
-                    'max_requests' => 2
+        new RequestLimitRuleset(
+            [
+                'www.test.de' => [
+                    [
+                        'max_requests' => 2
+                    ]
                 ]
-            ]
-        ],
+            ],
             'cache',
             'laravel',
             new Repository([
@@ -45,9 +40,8 @@ class MemcachedDriverTest extends TestCase
         );
     }
 
-    /** @test
-     */
-    public function requests_are_cached() : void
+    /** @test */
+    public function requests_are_cached(): void
     {
         $this->_memcachedRequired();
 
@@ -64,13 +58,14 @@ class MemcachedDriverTest extends TestCase
         $memcached->flush();
 
         $host = 'www.test.de';
-        $ruleset = new RequestLimitRuleset([
-            $host => [
-                [
-                    'max_requests' => 2
+        $ruleset = new RequestLimitRuleset(
+            [
+                $host => [
+                    [
+                        'max_requests' => 2
+                    ]
                 ]
-            ]
-        ],
+            ],
             'cache',
             'laravel',
             new Repository([
@@ -86,7 +81,8 @@ class MemcachedDriverTest extends TestCase
                         ]
                     ]
                 ]
-            ]));
+            ])
+        );
         $throttle = new ThrottleMiddleware($ruleset);
         $stack = new MockHandler([new Response(200, [], null, '1'), new Response(200, [], null, '2'), new Response(200, [], null, '3')]);
         $client = new Client(['base_uri' => $host, 'handler' => $throttle->handle()($stack)]);
@@ -99,17 +95,15 @@ class MemcachedDriverTest extends TestCase
         static::assertEquals($responseTwo, $responseThree);
     }
 
-    private function _memcachedRequired() : void
+    private function _memcachedRequired(): void
     {
-        if (!class_exists('Memcached'))
-        {
-            self::markTestSkipped('Memcached is required for this test.');
+        if (!extension_loaded('memcached')) {
+            self::markTestSkipped('The memcached extension is required for this test.');
         }
     }
 
-    /** @test
-     */
-    public function throw_too_many_requests_when_nothing_in_cache() : void
+    /** @test */
+    public function throw_too_many_requests_when_nothing_in_cache(): void
     {
         $this->_memcachedRequired();
 
@@ -126,13 +120,14 @@ class MemcachedDriverTest extends TestCase
         $memcached->flush();
 
         $host = 'www.test.de';
-        $ruleset = new RequestLimitRuleset([
-            $host => [
-                [
-                    'max_requests' => 0
+        $ruleset = new RequestLimitRuleset(
+            [
+                $host => [
+                    [
+                        'max_requests' => 0
+                    ]
                 ]
-            ]
-        ],
+            ],
             'cache',
             'laravel',
             new Repository([
@@ -142,7 +137,8 @@ class MemcachedDriverTest extends TestCase
                         'servers' => $servers
                     ]
                 ]
-            ]));
+            ])
+        );
         $throttle = new ThrottleMiddleware($ruleset);
         $stack = new MockHandler([new Response()]);
         $client = new Client(['base_uri' => $host, 'handler' => $throttle->handle()($stack)]);
@@ -150,5 +146,4 @@ class MemcachedDriverTest extends TestCase
         $this->expectException(TooManyRequestsHttpException::class);
         $client->request('GET', '/');
     }
-
 }
